util: check for error from AutoFile::fclose() in explicit calls
Explicit callers of `AutoFile::fclose()` should check whether
`fclose(3)` failed because this may be due to a failure to
flush data to disk. Such failures should be treated as failures to
`fwrite(3)`.

Also terminate the program if `AutoFile::~AutoFile()` is unable
to close the file because that is a file write failure which
cannot be handled inside the destructor or reported to the caller.

Github-Pull: #29307
Rebased-From: f9cf6b516b796461feb059729d2ab3737ad01ccc